Relocate AotProcessingHook
Relocate `AotProcessingHook` to be an static class of `AotProcessor`.
